第11讲 StreamConfigurationMap实战一

+++ GeekCamera2/app/src/main/java/com/deepinout/geekcamera/preview/VideoProfile.java
@@ -28,7 +28,7 @@ public class VideoProfile {
     public int videoSource;
     public int videoCodec;
     public int videoFrameRate;
-    public double videoCaptureRate;
+    public int videoCaptureRate;
     public int videoBitRate;
     public int videoFrameHeight;
     public int videoFrameWidth;
